+++ src/pyopenvba/powerquery/__init__.py
@@ -22,6 +22,7 @@ from pyopenvba.powerquery._files import pull_queries, push_queries
 from pyopenvba.powerquery._mashup import Mashup
 from pyopenvba.powerquery._metadata import Entry, Item, Metadata, QueryGroup
 from pyopenvba.powerquery._package import Package
+from pyopenvba.powerquery._refresh import RefreshSettings
 from pyopenvba.powerquery._section import Section, quote_name, unquote_name
 from pyopenvba.powerquery.workbook import (
     LOAD_CONNECTION_ONLY,
@@ -43,6 +44,7 @@ __all__ = [
     "PowerQuery",
     "PowerQueryWorkbook",
     "QueryGroup",
+    "RefreshSettings",
     "Section",
     "quote_name",
     "unquote_name",